move sfn test utils and fixtures to localstack.testing
