+++ tests/aws/services/stepfunctions/v2/activities/test_activities.py
@@ -3,12 +3,12 @@ import json
 from localstack_snapshot.snapshots.transformer import RegexTransformer
 
 from localstack.testing.pytest import markers
+from localstack.testing.pytest.stepfunctions.utils import (
+    create_and_record_execution,
+)
 from localstack.utils.strings import short_uid
 from tests.aws.services.stepfunctions.templates.activities.activity_templates import (
     ActivityTemplate,
-)
-from tests.aws.services.stepfunctions.utils import (
-    create_and_record_execution,
 )
