+++ tests/aws/services/stepfunctions/v2/base/test_base.py
@@ -7,13 +7,13 @@ from localstack_snapshot.snapshots.transformer import JsonpathTransformer
 
 from localstack.testing.aws.util import is_aws_cloud
 from localstack.testing.pytest import markers
-from localstack.utils.strings import short_uid
-from tests.aws.services.stepfunctions.templates.base.base_templates import BaseTemplate
-from tests.aws.services.stepfunctions.utils import (
+from localstack.testing.pytest.stepfunctions.utils import (
     await_execution_success,
     create_and_record_events,
     create_and_record_execution,
 )
+from localstack.utils.strings import short_uid
+from tests.aws.services.stepfunctions.templates.base.base_templates import BaseTemplate
 
 
 @markers.snapshot.skip_snapshot_verify(paths=["$..loggingConfiguration", "$..tracingConfiguration"])
@@ -165,6 +165,7 @@ class TestSnfBase:
         self,
         create_iam_role_for_sfn,
         create_state_machine,
+        events_to_sqs_queue,
         sfn_events_to_sqs_queue,
         aws_client,
         sfn_snapshot,
@@ -188,7 +189,6 @@ class TestSnfBase:
         self,
         create_iam_role_for_sfn,
         create_state_machine,
-        sfn_events_to_sqs_queue,
         aws_client,
         sfn_snapshot,
     ):
