+++ tests/aws/services/stepfunctions/v2/choice_operators/utils.py
@@ -4,11 +4,11 @@ from typing import Any, Final
 from localstack_snapshot.snapshots.transformer import RegexTransformer
 
 from localstack.services.stepfunctions.asl.utils.json_path import JSONPathUtils
+from localstack.testing.pytest.stepfunctions.utils import await_execution_success
 from localstack.utils.strings import short_uid
 from tests.aws.services.stepfunctions.templates.choiceoperators.choice_operators_templates import (
     ChoiceOperatorTemplate as COT,
 )
-from tests.aws.services.stepfunctions.utils import await_execution_success
 
 TYPE_COMPARISONS: Final[list[tuple[Any, bool]]] = [
     (None, True),  # 0
