+++ tests/aws/services/stepfunctions/v2/comments/test_comments.py
@@ -3,6 +3,7 @@ import json
 from localstack_snapshot.snapshots.transformer import RegexTransformer
 
 from localstack.testing.pytest import markers
+from localstack.testing.pytest.stepfunctions.utils import create_and_record_execution
 from localstack.utils.strings import short_uid
 from tests.aws.services.stepfunctions.templates.comment.comment_templates import (
     CommentTemplates as CT,
@@ -10,7 +11,6 @@ from tests.aws.services.stepfunctions.templates.comment.comment_templates import
 from tests.aws.services.stepfunctions.templates.services.services_templates import (
     ServicesTemplates as ST,
 )
-from tests.aws.services.stepfunctions.utils import create_and_record_execution
 
 
 @markers.snapshot.skip_snapshot_verify(paths=["$..loggingConfiguration", "$..tracingConfiguration"])
