+++ tests/aws/services/stepfunctions/v2/error_handling/test_task_service_sqs.py
@@ -5,6 +5,9 @@ from localstack_snapshot.snapshots.transformer import JsonpathTransformer, Regex
 
 from localstack.testing.aws.util import is_aws_cloud
 from localstack.testing.pytest import markers
+from localstack.testing.pytest.stepfunctions.utils import (
+    create_and_record_execution,
+)
 from localstack.utils.strings import short_uid
 from tests.aws.services.stepfunctions.templates.errorhandling.error_handling_templates import (
     ErrorHandlingTemplate as EHT,
@@ -12,7 +15,6 @@ from tests.aws.services.stepfunctions.templates.errorhandling.error_handling_tem
 from tests.aws.services.stepfunctions.templates.services.services_templates import (
     ServicesTemplates as ST,
 )
-from tests.aws.services.stepfunctions.utils import create_and_record_execution
 
 
 @markers.snapshot.skip_snapshot_verify(
