+++ tests/aws/services/stepfunctions/v2/error_handling/utils.py
@@ -1,7 +1,7 @@
 from localstack_snapshot.snapshots.transformer import RegexTransformer
 
+from localstack.testing.pytest.stepfunctions.utils import await_execution_success
 from localstack.utils.strings import short_uid
-from tests.aws.services.stepfunctions.utils import await_execution_success
 
 
 @staticmethod
