+++ tests/aws/services/stepfunctions/v2/services/test_apigetway_task_service.py
@@ -7,6 +7,9 @@ from localstack import config
 from localstack.aws.api.lambda_ import Runtime
 from localstack.testing.aws.util import is_aws_cloud
 from localstack.testing.pytest import markers
+from localstack.testing.pytest.stepfunctions.utils import (
+    create_and_record_execution,
+)
 from localstack.utils.aws import arns, aws_stack
 from localstack.utils.strings import short_uid
 from tests.aws.services.apigateway.apigateway_fixtures import create_rest_resource
@@ -14,7 +17,6 @@ from tests.aws.services.apigateway.conftest import APIGATEWAY_ASSUME_ROLE_POLICY
 from tests.aws.services.stepfunctions.templates.services.services_templates import (
     ServicesTemplates as ST,
 )
-from tests.aws.services.stepfunctions.utils import create_and_record_execution
 
 
 @markers.snapshot.skip_snapshot_verify(
