+++ tests/aws/services/stepfunctions/v2/services/test_events_task_service.py
@@ -4,11 +4,14 @@ import pytest
 from localstack_snapshot.snapshots.transformer import RegexTransformer
 
 from localstack.testing.pytest import markers
+from localstack.testing.pytest.stepfunctions.utils import (
+    create_and_record_execution,
+    record_sqs_events,
+)
 from localstack.utils.strings import short_uid
 from tests.aws.services.stepfunctions.templates.services.services_templates import (
     ServicesTemplates as ST,
 )
-from tests.aws.services.stepfunctions.utils import create_and_record_execution, record_sqs_events
 
 
 @markers.snapshot.skip_snapshot_verify(
