+++ tests/aws/services/stepfunctions/v2/services/test_lambda_task_service.py
@@ -6,11 +6,13 @@ from localstack_snapshot.snapshots.transformer import JsonpathTransformer, Regex
 from localstack.aws.api.lambda_ import LogType
 from localstack.testing.aws.util import is_aws_cloud
 from localstack.testing.pytest import markers
+from localstack.testing.pytest.stepfunctions.utils import (
+    create_and_record_execution,
+)
 from localstack.utils.strings import short_uid
 from tests.aws.services.stepfunctions.templates.services.services_templates import (
     ServicesTemplates as ST,
 )
-from tests.aws.services.stepfunctions.utils import create_and_record_execution
 
 
 @markers.snapshot.skip_snapshot_verify(
